Chap03/PythonVisionAPI/PythonVisionAPI/requirements.txt
absl-py==0.6.1
astor==0.7.1
certifi==2018.10.15
chardet==3.0.4
Click==7.0
cycler==0.10.0
DateTime==4.3
dlib==19.17.0
editdistance==0.5.3
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.0.3
gast==0.2.0
grpcio==1.17.1
h5py==2.9.0
idna==2.7
itsdangerous==1.1.0
jieba==0.39
Jinja2==2.10
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.1.0
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.3
nltk==3.4.1
numpy==1.15.4
openalpr==1.0.12
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
pip==19.1.1
pprint==0.1
protobuf==3.6.1
PyAudio==0.2.11
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.20.0
scikit-learn==0.21.2
scipy==1.2.1
setuptools==39.0.1
six==1.12.0
SpeechRecognition==3.8.1
tensorboard==1.12.1
tensorflow==1.12.0
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
wheel==0.32.3
zope.interface==4.6.0
